backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
celery
redis
httpx
python-dotenv
alembic
